refactor(statistics): Lotto Result 클래스 위치 변경
