+++ src/main/java/lotto/domain/ticket/Lotto.java
@@ -9,7 +9,7 @@ import static lotto.domain.ticket.LottoNumberConfig.NUMBER_COUNT;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.domain.dto.LottoResult;
+import lotto.domain.statistics.LottoResult;
 import lotto.domain.winningnumbers.WinningNumbers;
 
 public class Lotto {
